Fix compiler warnings

+++ DashControl/AppDelegate.m
@@ -286,7 +286,7 @@ static NSString* NSStringFromQueryParameters(NSDictionary* queryParameters)
     NSURLSessionDataTask* task = [session dataTaskWithRequest:request completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
         if (error == nil) {
             // Success
-            NSLog(@"URL Session Task Succeeded: HTTP %ld", ((NSHTTPURLResponse*)response).statusCode);
+            NSLog(@"URL Session Task Succeeded: HTTP %ld", (long)((NSHTTPURLResponse*)response).statusCode);
         }
         else {
             // Failure
